Fix 64-bit cast #jira UE-30132 #ue4 #android
[CL 2961278 by Chris Babcock in 4.12 branch]

+++ Engine/Source/Runtime/Core/Private/Android/AndroidMisc.cpp
@@ -648,7 +648,7 @@ void FAndroidMisc::SetCrashHandler(void (* CrashHandler)(const FGenericCrashCont
 	FMemory::Memzero(&PrevActions, sizeof(PrevActions));
 
 	// Passing -1 will leave these restored and won't trap them
-	if ((int)CrashHandler == -1)
+	if ((int64)CrashHandler == -1)
 	{
 		return;
 	}
